improve readme; applied code style fixes

+++ src/test/java/br/unb/cic/tdp/permutation/CycleTest.java
@@ -41,22 +41,22 @@ class CycleTest {
 
     @Test
     void image() {
-        assertEquals(1, cycle.image((int) 0));
+        assertEquals(1, cycle.image(0));
     }
 
     @Test
     void pow() {
-        assertEquals(2, cycle.pow((int) 0, 2));
+        assertEquals(2, cycle.pow(0, 2));
     }
 
     @Test
     void getK() {
-        assertEquals(2, cycle.getK((int) 0, (int) 2));
+        assertEquals(2, cycle.getK(0, 2));
     }
 
     @Test
     void getStartingBy() {
-        assertArrayEquals(new int[] {2, 3, 4, 5, 0, 1}, cycle.startingBy((int) 2).getSymbols());
+        assertArrayEquals(new int[] {2, 3, 4, 5, 0, 1}, cycle.startingBy(2).getSymbols());
     }
 
     @Test
@@ -67,13 +67,13 @@ class CycleTest {
 
     @Test
     void indexOf() {
-        assertEquals(3, cycle.indexOf((int) 3));
+        assertEquals(3, cycle.indexOf(3));
     }
 
     @Test
     void contains() {
-        assertTrue(cycle.contains((int) 3));
-        assertFalse(cycle.contains((int) 6));
+        assertTrue(cycle.contains(3));
+        assertFalse(cycle.contains(6));
     }
 
     @Test
